Avoid implementation-defined narrowing in correctionLaunch() (#1579)
* Avoid implementation-defined narrowing in correctionLaunch()

correctionLaunch() accumulated NO_FUEL_CORRECTION (100) + lnchFuelAdd
(tune-editable -40..+80) in an int8_t, which can hold a computed
value of 180 only via an implementation-defined conversion on
assignment (well-defined as two's-complement wraparound on this
project's GCC-based toolchains, but not something the standard
guarantees generally). Widened the accumulator to int16_t so the
value is correct by construction rather than by relying on a specific
narrowing-conversion behaviour; the final (uint8_t) cast is now a
plain, always-safe conversion since the result is always in [60,180].

Fixes speeduino#1571


* Add correctionLaunch tests at the tune min/max extremes

Covers the int16_t accumulator change: lnchFuelAdd = +80 (max) yields
180 and lnchFuelAdd = -40 (min) yields 60. The +80 case is the one
that overflowed the previous int8_t accumulator; this locks in the
correct result and covers the changed line.


* Address review: comment explaining the int16_t accumulator

Per @adbancroft's review on #1579.


---------

+++ speeduino/corrections.cpp
@@ -568,10 +568,13 @@ This simple check applies the extra fuel if we're currently launching
 */
 TESTABLE_INLINE_STATIC uint8_t correctionLaunch(void)
 {
-  int8_t correction = (int8_t)NO_FUEL_CORRECTION;
+  // Accumulate in int16_t: NO_FUEL_CORRECTION (100) + lnchFuelAdd (tune range
+  // -40..+80) can reach 180, which overflows int8_t. The result is always within
+  // [60, 180], so the final narrowing to uint8_t is safe.
+  int16_t correction = (int16_t)NO_FUEL_CORRECTION;
   if (currentStatus.launchingHard || currentStatus.launchingSoft) {
     correction = correction + configPage6.lnchFuelAdd;
-  } 
+  }
   return (uint8_t)correction;
 }
